Move const to fix import boundary errors
