+++ packages/react-ui/src/app/shared/oauth-channel-prefix.ts
@@ -1,0 +1,1 @@
+export const OAUTH_CHANNEL_PREFIX = 'oauth2-redirect-';
